requirements.txt
anyio==3.6.2
aspose-cells==23.1.0
aspose-words==23.1.0
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
distro==1.8.0
executing==1.2.0
fastapi==0.89.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
future==0.18.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
JPype1==1.4.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.28
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
Pygments==2.14.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.22.0
tabula-py==2.6.0
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Wand==0.6.11
wcwidth==0.2.6
Werkzeug==2.3.6
